School Project
This is the beginning of my school project. This project is a movie
ticket reservation system. This is a multi-user plugin. That is users
can buy tickets after logging in and then buy the ticket. After they
login again the next time they will be able to see their tickets and do
the required jobs.